handle empty source folder better

+++ ECMA2Yaml/ECMA2Yaml/Nuget/Run.ps1
@@ -49,7 +49,7 @@ if (-not $publicGitUrl.EndsWith("/"))
 $jobs = $ParameterDictionary.docset.docsetInfo.ECMA2Yaml
 if (!$jobs)
 {
-	$jobs = $ParameterDictionary.environment.publishConfigContent.ECMA2Yaml
+    $jobs = $ParameterDictionary.environment.publishConfigContent.ECMA2Yaml
 }
 if ($jobs -isnot [system.array])
 {
@@ -60,41 +60,45 @@ foreach($ecmaConfig in $jobs)
     $ecmaXmlGitUrlBase = $publicGitUrl + "blob/" + $publicBranch
     echo "Using $ecmaXmlGitUrlBase as url base"
     $ecmaSourceXmlFolder = Join-Path $repositoryRoot $ecmaConfig.SourceXmlFolder
+    if (-not (Test-Path $ecmaSourceXmlFolder))
+    {
+        continue;
+    }
     $ecmaOutputYamlFolder = Join-Path $repositoryRoot $ecmaConfig.OutputYamlFolder
     $allArgs = @("-s", "$ecmaSourceXmlFolder", "-o", "$ecmaOutputYamlFolder", "-l", "$logFilePath", "-p", """$repositoryRoot=>$ecmaXmlGitUrlBase""");
-	
-	$fallbackRepoRoot = Join-Path $repositoryRoot _repo.en-us
-	$ecmaFallbackSourceXmlFolder = Join-Path $fallbackRepoRoot $ecmaConfig.SourceXmlFolder
-	$fallbackRepo = $null
-	foreach($repo in $ParameterDictionary.environment.publishConfigContent.dependent_repositories)
-	{
-		if ($repo.path_to_root -eq "_repo.en-us")
-		{
-			$fallbackRepo = $repo
-		}
-	}
-	if ($fallbackRepo -and (Test-Path $ecmaFallbackSourceXmlFolder))
-	{
-		$allArgs += "-fs";
-		$allArgs += "$ecmaFallbackSourceXmlFolder";
-		$fallbackGitUrlBase = $fallbackRepo.url
-		if (-not $fallbackGitUrlBase.EndsWith("/"))
-		{
-			$fallbackGitUrlBase += "/"
-		}
-		$fallbackGitUrlBase = $fallbackGitUrlBase + "blob/" + $fallbackRepo.branch;
-		$allArgs +=  "-fp"
-		$allArgs +=  """$fallbackRepoRoot=>$fallbackGitUrlBase"""
-
-		if ([string]::IsNullOrEmpty($ParameterDictionary.environment.skipPublishFilePath)) {
-			$ParameterDictionary.environment.skipPublishFilePath = Join-Path $logOutputFolder "skip-publish-file-path.json"
-		}
-		$skipPublishFilePath = $ParameterDictionary.environment.skipPublishFilePath;
-		$allArgs +=  "-skipPublishFilePath"
-		$allArgs +=  "$skipPublishFilePath"
-	}
     
-	if ($ecmaConfig.Flatten)
+    $fallbackRepoRoot = Join-Path $repositoryRoot _repo.en-us
+    $ecmaFallbackSourceXmlFolder = Join-Path $fallbackRepoRoot $ecmaConfig.SourceXmlFolder
+    $fallbackRepo = $null
+    foreach($repo in $ParameterDictionary.environment.publishConfigContent.dependent_repositories)
+    {
+        if ($repo.path_to_root -eq "_repo.en-us")
+        {
+            $fallbackRepo = $repo
+        }
+    }
+    if ($fallbackRepo -and (Test-Path $ecmaFallbackSourceXmlFolder))
+    {
+        $allArgs += "-fs";
+        $allArgs += "$ecmaFallbackSourceXmlFolder";
+        $fallbackGitUrlBase = $fallbackRepo.url
+        if (-not $fallbackGitUrlBase.EndsWith("/"))
+        {
+            $fallbackGitUrlBase += "/"
+        }
+        $fallbackGitUrlBase = $fallbackGitUrlBase + "blob/" + $fallbackRepo.branch;
+        $allArgs +=  "-fp"
+        $allArgs +=  """$fallbackRepoRoot=>$fallbackGitUrlBase"""
+
+        if ([string]::IsNullOrEmpty($ParameterDictionary.environment.skipPublishFilePath)) {
+            $ParameterDictionary.environment.skipPublishFilePath = Join-Path $logOutputFolder "skip-publish-file-path.json"
+        }
+        $skipPublishFilePath = $ParameterDictionary.environment.skipPublishFilePath;
+        $allArgs +=  "-skipPublishFilePath"
+        $allArgs +=  "$skipPublishFilePath"
+    }
+    
+    if ($ecmaConfig.Flatten)
     {
         $allArgs += "-f";
     }
@@ -105,28 +109,28 @@ foreach($ecmaConfig in $jobs)
     if (-not [string]::IsNullOrEmpty($ecmaConfig.SourceMetadataFolder))
     {
         $ecmaSourceMetadataFolder = Join-Path $repositoryRoot $ecmaConfig.SourceMetadataFolder
-		if (Test-Path $ecmaSourceMetadataFolder)
-		{
-			$allArgs += "-m";
+        if (Test-Path $ecmaSourceMetadataFolder)
+        {
+            $allArgs += "-m";
             $allArgs += "$ecmaSourceMetadataFolder";
-		}
+        }
     }
-	$changeListFile = $ParameterDictionary.context.changeListTsvFilePath;
-	if (-not [string]::IsNullOrEmpty($changeListFile) -and (Test-Path $changeListFile))
-	{
-		$newChangeList = $changeListFile -replace "\.tsv$",".mapped.tsv";
-		$allArgs += "-changeList";
+    $changeListFile = $ParameterDictionary.context.changeListTsvFilePath;
+    if (-not [string]::IsNullOrEmpty($changeListFile) -and (Test-Path $changeListFile))
+    {
+        $newChangeList = $changeListFile -replace "\.tsv$",".mapped.tsv";
+        $allArgs += "-changeList";
         $allArgs += "$changeListFile";
-		$ParameterDictionary.context.changeListTsvFilePath = $newChangeList
-	}
-	$userChangeListFile = $ParameterDictionary.context.userSpecifiedChangeListTsvFilePath;
-	if (-not [string]::IsNullOrEmpty($userChangeListFile) -and (Test-Path $userChangeListFile))
-	{
-		$newUserChangeList = $userChangeListFile -replace "\.tsv$",".mapped.tsv";
-		$allArgs += "-changeList";
+        $ParameterDictionary.context.changeListTsvFilePath = $newChangeList
+    }
+    $userChangeListFile = $ParameterDictionary.context.userSpecifiedChangeListTsvFilePath;
+    if (-not [string]::IsNullOrEmpty($userChangeListFile) -and (Test-Path $userChangeListFile))
+    {
+        $newUserChangeList = $userChangeListFile -replace "\.tsv$",".mapped.tsv";
+        $allArgs += "-changeList";
         $allArgs += "$userChangeListFile";
-		$ParameterDictionary.context.userSpecifiedChangeListTsvFilePath = $newUserChangeList
-	}
+        $ParameterDictionary.context.userSpecifiedChangeListTsvFilePath = $newUserChangeList
+    }
     $printAllArgs = [System.String]::Join(' ', $allArgs)
     $ecma2yamlExeFilePath = Join-Path $currentDir $ecma2yamlExeName
     echo "Executing $ecma2yamlExeFilePath $printAllArgs" | timestamp
@@ -139,11 +143,14 @@ foreach($ecmaConfig in $jobs)
     if (-not [string]::IsNullOrEmpty($ecmaConfig.id))
     {
         $tocPath = Join-Path $ecmaOutputYamlFolder "toc.yml"
-        $newTocPath = Join-Path $ecmaOutputYamlFolder $ecmaConfig.id
-        if (-not (Test-Path $newTocPath))
+        if (Test-Path $tocPath)
         {
-            New-Item -ItemType Directory -Force -Path $newTocPath
+            $newTocPath = Join-Path $ecmaOutputYamlFolder $ecmaConfig.id
+            if (-not (Test-Path $newTocPath))
+            {
+                New-Item -ItemType Directory -Force -Path $newTocPath
+            }
+            Move-Item $tocPath $newTocPath
         }
-        Move-Item $tocPath $newTocPath
     }
 }
